# [test-only] JMH microbenchmarks