--- a/dialogue-jmh/build.gradle
+++ b/dialogue-jmh/build.gradle
@@ -3,12 +3,17 @@ versionsLock {
 }
 
 dependencies {
-    implementation project(':dialogue-core')
-    implementation project(':dialogue-test-common')
-    implementation 'com.google.guava:guava'
-    implementation 'org.openjdk.jmh:jmh-core'
-    implementation 'com.palantir.safethreadlocalrandom:safe-thread-local-random'
     implementation 'com.github.ben-manes.caffeine:caffeine'
+    implementation 'com.google.guava:guava'
+    implementation 'com.palantir.conjure.java.runtime:client-config'
+    implementation 'com.palantir.safe-logging:preconditions'
+    implementation 'com.palantir.safe-logging:safe-logging'
+    implementation 'com.palantir.safethreadlocalrandom:safe-thread-local-random'
+    implementation 'com.palantir.tritium:tritium-registry'
+    implementation 'org.openjdk.jmh:jmh-core'
+    implementation project(':dialogue-core')
+    implementation project(':dialogue-target')
+    implementation project(':dialogue-test-common')
 
     runtimeOnly 'org.apache.logging.log4j:log4j-slf4j-impl'
     runtimeOnly 'org.apache.logging.log4j:log4j-core'
@@ -39,3 +44,4 @@ task jmh(type: JavaExec) {
 tasks.withType(JavaCompile) {
     options.errorprone.errorproneArgs += [ '-Xep:HidingField:OFF',  '-Xep:Slf4jLogsafeArgs:OFF', '-Xep:NullAway:OFF' ]
 }
+tasks.checkImplicitDependenciesMain.enabled = false // not published, don't care