--- a/dialogue-jmh/src/main/java/com/palantir/dialogue/core/DialogueChannelBenchmark.java
+++ b/dialogue-jmh/src/main/java/com/palantir/dialogue/core/DialogueChannelBenchmark.java
@@ -1,0 +1,40 @@
+/*
+ * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.dialogue.core;
+
+import java.util.concurrent.TimeUnit;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.Fork;
+import org.openjdk.jmh.annotations.Measurement;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Warmup;
+
+@Measurement(iterations = 3, time = 3)
+@Warmup(iterations = 2, time = 5)
+@Fork(value = 1)
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+public class DialogueChannelBenchmark {
+
+    @Benchmark
+    public int testMethod() {
+        int a = 1;
+        int b = 2;
+        int sum = a + b;
+
+        return sum;
+    }
+}