… dialogue-jmh/src/main/resources/log4j2.xml
… versions.lock
… versions.props