## Before this PR

We're starting to land PRs that fiddle with small optimizations, without an easy way to demonstrate the before/after improvements. This runs the risk of us doing big refactors that aren't actually helpful.

## After this PR
==COMMIT_MSG==
A new `:dialogue-jmh` project allows running microbenchmarks from IntellIJ or from the command line using `./gradlew jmh`
==COMMIT_MSG==

Initial results show that PIN_UNTIL_ERROR is ~2x faster than BALANCED
```
Benchmark                           (headerDriven)  (numChannels)  (selectionStrategy)   Mode  Cnt     Score     Error   Units
NodeSelectionBenchmark.postRequest            true              2      PIN_UNTIL_ERROR  thrpt   12   467.786 ±  28.286  ops/ms
NodeSelectionBenchmark.postRequest            true              2          ROUND_ROBIN  thrpt   12  3616.095 ± 201.541  ops/ms
NodeSelectionBenchmark.postRequest            true              8      PIN_UNTIL_ERROR  thrpt   12   141.798 ±  25.086  ops/ms
NodeSelectionBenchmark.postRequest            true              8          ROUND_ROBIN  thrpt   12  1824.192 ±  74.748  ops/ms
NodeSelectionBenchmark.postRequest           false              2      PIN_UNTIL_ERROR  thrpt   12  8020.820 ± 221.295  ops/ms
NodeSelectionBenchmark.postRequest           false              2          ROUND_ROBIN  thrpt   12  4672.392 ± 101.247  ops/ms
NodeSelectionBenchmark.postRequest           false              8      PIN_UNTIL_ERROR  thrpt   12  6694.895 ± 195.309  ops/ms
NodeSelectionBenchmark.postRequest           false              8          ROUND_ROBIN  thrpt   12  2031.100 ±  40.723  ops/ms
```

Also i've got one that just runs our generated code against undertow.

```
Benchmark                             Mode  Cnt      Score      Error  Units
EndToEndBenchmark.dialogueBlocking   thrpt   16  24013.938 ± 2141.187  ops/s
EndToEndBenchmark.rawApacheBlocking  thrpt   16  29994.941 ± 3443.332  ops/s
```

## Possible downsides?
- it doesn't run on CI, so could decay. In practise I don't really care because we can easily fix it when we want to run it again.
- we don't have an easy way to graph changes over O(months)